python-promtail-loki-prom-grafana-demo/app/requirements.txt
flask
prometheus-client
requests

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

deprecated